fix: change data type from MONEY to REAL for price in books table

+++ sql_crud/0-create_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE books (
     title TEXT NOT NULL,
     author TEXT NOT NULL,
     genre TEXT NOT NULL,
-    price MONEY NOT NULL,
+    price REAL NOT NULL,
     stock INTEGER,
     published_year INTEGER NOT NULL
 );
